requirements.txt
discord.py>=2.6.0
python-dotenv
aiohttp
mcstatus
